refactor(web): formalize e2e mock api budgets
