+++ web/file-budgets.config.mjs
@@ -5,6 +5,7 @@ export const fileBudgetLimits = {
   featureTest: { soft: 300, hard: 650 },
   featureStateModule: { soft: 250, hard: 500 },
   featureModule: { soft: 200, hard: 325 },
+  testingSupportModule: { soft: 350, hard: 650 },
   layoutComponent: { soft: 200, hard: 300 },
   uiPrimitive: { soft: 150, hard: 250 },
 }
@@ -31,6 +32,10 @@ function isFeatureComponent(filePath) {
 
 function isFeatureModule(filePath) {
   return /^src\/lib\/features\/.+\.(ts|js|mjs|cjs)$/.test(filePath)
+}
+
+function isTestingSupportModule(filePath) {
+  return /^src\/lib\/testing\/.+\.(ts|js|mjs|cjs)$/.test(filePath)
 }
 
 function isLayoutComponent(filePath) {
@@ -80,6 +85,12 @@ export const fileBudgetRules = [
       !isFeatureStateModule(filePath),
     softLimit: fileBudgetLimits.featureModule.soft,
     hardLimit: fileBudgetLimits.featureModule.hard,
+  },
+  {
+    name: 'Testing support modules',
+    match: isTestingSupportModule,
+    softLimit: fileBudgetLimits.testingSupportModule.soft,
+    hardLimit: fileBudgetLimits.testingSupportModule.hard,
   },
   {
     name: 'Layout components',
